report lag metrics only on commits events, report execution time on real measurements
