+++ src/main/java/com/zendesk/maxwell/replication/AbstractReplicator.java
@@ -33,7 +33,6 @@ public abstract class AbstractReplicator extends RunLoopProcess implements Repli
 
 	private final Counter rowCounter;
 	private final Meter rowMeter;
-	private Long replicationLag = 0L;
 
 	public AbstractReplicator(
 		String clientID,
@@ -49,18 +48,6 @@ public abstract class AbstractReplicator extends RunLoopProcess implements Repli
 		this.producer = producer;
 		this.lastHeartbeatPosition = initialPosition;
 
-		final AbstractReplicator self = this;
-
-		String lagGaugeName = metrics.metricName("replication", "lag");
-		metrics.getRegistry().register(
-			lagGaugeName,
-			new Gauge<Long>() {
-				@Override
-				public Long getValue() {
-					return self.replicationLag;
-				}
-			}
-		);
 		rowCounter = metrics.getRegistry().counter(
 			metrics.metricName("row", "count")
 		);
@@ -161,7 +148,6 @@ public abstract class AbstractReplicator extends RunLoopProcess implements Repli
 
 		rowCounter.inc();
 		rowMeter.mark();
-		replicationLag = System.currentTimeMillis() - row.getTimestampMillis();
 
 		processRow(row);
 	}
